fix: 참가자 수를 DB에 저장하는 방식에서 redis를 사용하는 방식으로 변경

+++ src/main/java/com/hong/ForPaw/controller/DTO/SearchResponse.java
@@ -28,7 +28,7 @@ public class SearchResponse {
     public record GroupDTO(Long id,
                            String name,
                            String description,
-                           Integer participationNum,
+                           Long participationNum,
                            String category,
                            String region,
                            String subRegion,
